Fix handling of screen changes
Move connecting to QWindow::screenChanged to our widget's showEvent.
This is necessary as the window may not be fully constructed when our
constructor executes.

More importantly, because this is part of a dock widget, the window
actually changes when the containing dock widget is docked or undocked.
Always resetting the connection whenever a showEvent happens ensures
that we are always monitoring the correct window.