(WIP) Update manual
Update the user manual to document some of the new (and not so new)
features that have been added. Also, get rid of double blank lines, as
these were not being used consistently and it's easier to remember and
potentially enforce their omission.

This still needs work (note the "TODO"s), but I am less familiar with
the parts that still need polish and am hoping that one of the other
developers can follow up on this.