Mark camera path as modified after adding points.
This was not needed prior to VTK 8, but now it is required.
This change fixes two bugs.  First, the camera path has not been
rendering since upgrading to VTK 8.  Second, the scale of the cameras
has been off because it uses the bounds on the camera path to compute
the base scale and the camera path was undefined.