Add High-DPI support
Add additional icons for (2x) High-DPI displays. According to the Qt
docs, for 2x scaling, the icon loading guts should "discover" these
automatically as long as they are present, with code changes required
besides adding the correctly named additional resource files.

Also explicitly enable AA_EnableHighDpiScaling, which for some
mysterious reason is not on by default (causing Qt apps to mostly, but
not fully, respond to display scaling) and is actually needed for
high-DPI icons to be selected by QIcon, as well as for some other things
to be done correctly (e.g. there are some style glitches without it).