Split setRobustROI into separate compute and set functions
The compute function is now used to estimate the scene scale if no ROI is
defined, but does so without actually saving the ROI.  This way we only define
the ROI when asked to do so.  This fixes a bug in which the ROI was created
after translating landmarks and reestimating scale, and then the ROI was also
translated resulting in a double translate.