Avoid camera flickering when setting the same frame that is already active.
The camera should only be set to null if the camera is not found.
If the camera is not changed we should not update it, but also not set it
to null.  This was causing the camera to disappear in some cases when
the scene was updated without changing the active camera.