Fixed bug that shifted geo-origin but not everything else
Shifting the geo-origin was happening after every intermediate step of the
tools. This would have been fine except that the tools maintain their own
internal copy of the landmarks, cameras, etc. but not geo-origin.  So each
time the geo-origin was shifted the cameras and landmarks were reset so the
geo-origin kept drifter further away.  The solution is to only shift the
geo-origin when a tool is complete and gives up its state.