Fix a bug in which camera frustums did not render.
This was a long standing bug in which camera frustums failed to render only
on Linux and only in optimized builds.  It was not a problem in Debug builds
or any Windows builds.  The problem was that the image size was being set to
zero which produced invalid (nan) geometry.  This happened because the auto
keyword was used in conjuction with the result of an Eigen expression,
which causes the auto to store an Eigen expression that may not yet be evaluated.
We need to use a concrete type (vector_2d) to force Eigen to actually evaluate
the expression before accessing the results.