Add images for apply icon
Generate and add PNG images for the "apply" icon. These are intended to
be used for the action to update the georegistration based on registered
ground control points. It is a little generic for this purpose, but has
the advantage that the icon graphics already existed, and just were not
in use before now.