Fixed an issue with GCPs behind the camera projecting to random locations
When a GCP was moved in the world view from in front of the active camera to
behind it, the hidden GCP would reappear at random locations in the camera
view.  This fix sets the location of an invalid projections to infinite values
so that they are not rendered.  The same is true for the ruler widget end
points except additional logic is needed to not draw the line if either end
point is at infinity.