Constrain ruler endpoint motion using coordinates from the other endpoint
This change modifies how the interaction with the endpoint is performed.
Z axis constrained motion would ensure that the X and Y coordinates of
both the endpoints are same (i.e. ruler is a vertical line), whereas XY
plane constrained motion ensures that the Z coordinate of the both the
endpoints is the same.