Avoid decelerated motion when moving towards the near and far planes
The original change would slow constrained motion when moving towards
the near or far clipping planes with the mouse cursor. This was because
the mouse position was projected on to the focal plane and then
constrained. The new approach directly projects the mouse position onto
the constraint plane. In the case of XY plane motion, the mouse position
is projected onto the XY plane. In the case of Z axis constraint, the
mouse position is projected onto a plane passing through the current
position and parallel to the Z axis.