Fix a bug in code that checks image sizes.
When the image size check was added it didn't account for the fact that
the depth image can be cropped but the color image is not cropped.
This revised check ensures that all the depth pixels are within the bounds
of the color images, but sizes need not be equal.