Speed up video playback by limiting seeking.
Playing back video in TeleSculptor was slower than necessary because advancing
the video is always done by sending a message to go to a specific frame number.
Previously this meant that seek was called for each frame change, even if the
frame number was just stepping to the next frame.  Seek calls step the video
back to the previous key frame and then forward again to reach the target
frame.  It is more efficient to first check if the video is already just a few
frames behind the target and then advance the video in that case instead of
seeking.