Update Metadata fields using current metadata
This branch is the start of change that moves away from pre-scanning the
entire video into a metadata map.  For the MetadataView it is a relatively
simple change to always update to the current frame metadata.  The main
difference is that the full set of metadata keys is not known in advance.
The set of keys must expand as each new frame is visited if a new key
is encountered.