Refactored code to load all cameras from KRTD files
This code is now in a helper function.  Also it now checks that each
file exists before trying to open it.  This is much faster than trying
to open a file for each frame and catching exceptions for each missing
file.  As a result the warning message about failing to load a KRTD
now is only issued when a file exists but still fails to load.