Correct the way depth saving works for batch depth maps
Previously the Batch Compute Depth tool saved depth to disk on
every update.  This was fine when updates were only made once for
each frame, but the code was updated to show intermediate results.
Due to a race condition, sometimes an intermediate result was saved
over the final result.  This branch removes the old special case
code for saving batch depth maps and updates to use the newer
framework for saving intermediate results that was introduced with
the Run End-to-End tool.