Fix how WorldView::fusedMeshEnabled is emitted
Add missing `emit` annotation to one use of fusedMeshEnabled. Remove
connections to emit the signal created when loading a volume (one of
which was already redundant!) and instead emit the signal via
setVolumeVisible. This greatly simplifies things, and also fixes an
issue that directly calling the aforementioned method previously would
have bypassed the signal being emitted.