Refactor GCP helper for registration points
Refactor GroundControlPointsHelper in order to start preparing for this
class to also manage registration points. We need these to be managed by
a single class in order to have a centralized location for managing the
identifiers, since we eventually want to be able to add registration
points before they have an associated ground control point.

Also, refactor users so that the internal map is no longer directly
accessible, as the value type of said map is now a private type. Certain
batch operations (such as resetting the GCP view) may be somewhat less
efficient as a result, but efficiency on the whole should be similar,
and the operations most impacted are not believed to be major
bottlenecks.