Consistently use frame_id_t
Modify most uses of frame identifiers to consistently use frame_id_t as
the data type for the same. This isn't perfect, as the UI still must use
int in some places, but overall this should significantly improve
consistency and correctness.

Additionally, rename WorldView::setActiveCamera to setActiveFrame to
better distinguish between the camera object and camera/frame id. (This
is also preparatory to adding a method to get the current id, which will
necessarily need a different name.)