More work on "registration points"
Start hooking up logic to actually create feature tracks for
registration points, and to manage them in the GCP model/view. Note that
updates (moving points after placement) are not hooked up yet, nor is
reshowing the points after changing what camera is active.

There are also some bugs with respect to selecting points that can cause
the tree view and world/camera view's notion of what is active to be out
of sync.

Note that this also introduces the notion that a "ground control point"
might consist solely of a feature track and not actually have an
associated world location.