Allow deselecting active seed
Modify vtkMaptkSeedWidget::HighlightActiveSeed to not change the active
seed if it is invalid. This is necessary to allow us to "clear" the seed
"selection", which is needed to properly support sparse correspondence
between camera registration points and ground control points. In
particular, this fixes selection in the world/camera views not matching
the selection in the list (tree) view, and makes it possible to add a
ground control point to an existing camera registration point.