Use selection, not current, for GCP
Modify GroundControlPointsView to use the selection (which should always
be a single item) as opposed to the current index as the "active" point.
While these will often be the same, there are some subtle instances in
which they aren't, such as when the list is non-empty but the selection
is empty, particularly if the selection has been manually cleared.