Fix active GCP bug
Modify GroundControlPointsWidget to suppress sending updates about the
active point when the widget is enabled. This should not be changing the
active point, but due to an implementation quirk (particularly, the VTK
widget can't have an active point when it is disabled, so we have to
track the active point ourselves, which means "restoring" the active
point when the widget is enabled causes a "spurious" change event), was
causing us to signal that the active point had changed.

This was effectively a non-issue before registration points, because we
didn't have the possibility that the active point should be valid, but
the widget wouldn't have that point. Now, however, we can have a CRP
with no GRP, or vice versa, which results in one of the widgets having
an invalid active point even though there is a valid active point.