Overhaul how point edit modes are enabled
Move logic to enable point editing widgets into the respective views,
purging it from GroundControlPointsHelper. This requires adding
additional API's to the views, but allows us to eliminate the need to
"lock out" editing registration points to edit ground control points
(which was previously needed as there was no API to tell the CameraView
to turn editing off).