Fix GCP selection issue
Modify GroundControlPointsWidget to not attempt to change the active
point if there are no points. This fixes a bug where trying to set the
active point to a non-existing point (which can happen when trying to
set the active GCP in the camera view when no camera is available) would
cause a change notification that would reset the active point, leading
to all sorts of issues.

Possibly we could arrange to not try to set the active point if we don't
expect the widget to be populated, but this approach is simpler.