Manage cameras ad-hoc, not in batch
Modify how we propagate camera updates to ColorizeSurfaceOptions to send
individual changes rather than sending the entire camera map as one
giant lump. Although this seems like it would be less efficient, the
"newly extra" logic in WorldView is all of the "mark dirty and update
later" form, and we would otherwise need to build up a map anyway, so
really this is just deepening the call stack for inserting cameras into
the map.

The benefit of this is that we can make individual updates without
having to completely rebuild the camera map. Also, it means that
MainWindowPrivate::updateCamera updates the ColorizeSurfaceOptions
cameras rather than having to remember to do this as a separate step.

Additionally, this lets us separate maintaining the cameras in
ColorizeSurfaceOptions and updating the WorldView camera representation,
which means we can (efficiently!) do so when a camera is computed, thus
fixing the bug that the representation was not being updated in this
case.