Return handle of new seed
Modify methods of GroundControlPointsWidget to return the handle
identifier of the newly added seed. Since GroundControlPointsHelper
often needs the handle to update its maps, this will, in at least some
cases, save having to guess that the new handle will be the same as the
old number-of-handles. Also, it is near-negligible cost to return this
since we already use the handle identifier while creating the seed.