Allow using system packages
Modify superbuild to add options to use system packages rather than
building everything in fletch. Some of these (Qt, VTK, OpenCV) are quite
large and using the system versions may save quite some time (and also
headaches, in the case of Qt, which can be rather temperamental).

Note that this support is fairly crude with the intent being that it is
up to the user to deal with any issues that come up with trying to use
system packages.