Avoid use of non-ASCII whitespace on Windows
Replace use of em-space for faking a right margin with (on Windows) two
regular spaces, as apparently Windows' Unicode coverage is so lame as to
lack even this basic non-ASCII character. Also, use a regular space
instead of a figure space to try to line up the numbers with and without
parentheses. Neither is perfect, but regular space actually seems to be
slightly less 'off'.