Initial ability to reconstruct from manual tracks
Introduce an action to attempt reconstruction using camera registration
points (i.e. manual feature tracks). This is partly implemented; it runs
and, in theory, extracts cameras. However, as I have not yet been able
to get it to run successfully, I am unable to verify the functionality.
Also, converting the reconstructed landmarks to GCPs is not yet
implemented (for similar reasons).