Hook up landmarks from reconstruction
Add ability to set the (world) location of a ground control point (which
may or may not already exist) to the helper. Use this to implement
accepting reconstructed world locations from reconstruction using manual
annotations, as added by the previous commit.

Although "real" reconstruction is not functioning as of this commit, the
logic has been tested by replacing the actual algorithm with creation of
random cameras and landmarks.