"Fix" rendering of single camera
Don't update the camera path if there is only one camera. Besides being
pointless (as we obviously cannot render a path with only one point),
for some mysterious reason trying to do so prevents the other actors
from rendering at all, resulting in not being able to render a single
camera.

This feels like a definite kludge, but seems to work...