Flatten documentation tree
Move documentation files out of 'manuals' subdirectory. The historic
reason for this subdirectory was that we might have other sorts of
documentation (e.g. internal developer documentation), but it doesn't
seem likely that will be the case any time soon. Also, having some
things up a level creates an artificial distinction which is not useful.

Files for the old in-application manual are left where they were, for
now (which is a useful distinction). Eventually that documentation is
expected to be incorporated into the RTD documentation, and the leftover
bits removed.