Start cleaning up RTD documentation
Move specification of the application version to conf.py, where it
belongs. Add KWIVER version to the same location. These changes will
give us a single place, at least within the RTD contents, where version
numbers and copyright year need to be kept up to date.

Update uses of the version number. Remove unneeded use of a replacement
directive to artificially insert a space. Improve formatting of file
paths. Add custom CSS for this. Start improving formatting of UI
elements referenced in the documentation.

Replace the large PNG logo with a slimmed-down version of the SVG, which
will render better across different screen resolutions. (We should do
this with the title logo, also, but not doing it right now.)

Make KWIVER reference in the root page a link.

Also, start reformatting documentation source to something more sane.
In particular, with reST, sentences should always begin a new line.
Sentences should also break at "logical boundaries" in order to keep
line lengths under control and minimize churn in case of changes (since
placement of breaks has no effect on the output). This only partly
covers those files which so far have been touched; others will receive
similar treatment as they are updated. Similarly, remove smart quotes
that we likely inherited from this documentation originating as a PDF;
some of them (e.g. appearing in text to be entered at a command prompt)
were outright wrong, and as for the rest, smart quotes are harder to
type in most plain-text editors such as are likely to be used to edit
the reST, and are not useful as Sphinx automatically generates smart
quotes.