Overhaul camera calibration documentation
Overhaul the "Manual Camera Calibration" section of the user
documentation. Add additional styling for menus, actions, shortcuts, and
inline mathematical notation. Add customizations to start allowing easy
referencing of of icons (for now, only as part of an action, but adding
a stand-alone role will be trivial), which was the main feature from the
old, non-RTD documentation system.

The updated "Manual Camera Calibration" section covers the same concepts
but has been significantly rewritten using the new features and to
improve grammar and tone.