Start reintegrating old user manual
Move the file containing the old user documentation into the main tree
for the RTD documentation. Update the customization to support the
'icon' role used by the same. Remove the old, no longer needed files for
generating the old manual, as the important parts of this are now
integrated into the RTD documentation.